major improvements to example, client and cloud
