+++ example/Makefile
@@ -1,5 +1,19 @@
-.PHONY: direct
+.PHONY: direct aws gcp azure web mount
 
 direct:
 	go run direct/main.go
 
+aws:
+	go run cloud/aws/main.go
+
+gcp:
+	go run cloud/gcp/main.go
+
+azure:
+	go run cloud/azure/main.go
+
+web:
+	go run client/web/main.go
+
+mount:
+	go run client/web/mount/main.go
