--- example/client/Makefile
@@ -1,10 +1,0 @@
-.PHONY: web mount-gin mount-http
-
-web:
-	go run web/main.go
-
-mount-gin:
-	go run mount/gin/main.go
-
-mount-http:
-	go run mount/http/main.go
