--- example/cloud/Makefile
@@ -1,10 +1,0 @@
-.PHONY: cloud-aws cloud-gcp cloud-azure
-
-cloud-aws:
-	go run cloud/aws/main.go
-
-cloud-gcp:
-	go run cloud/gcp/main.go
-
-cloud-azure:
-	go run cloud/azure/main.go
