+++ example/go.mod
@@ -1,30 +1,22 @@
-module buckttesting
+module example
 
 go 1.24.0
 
-require github.com/Rhaqim/buckt v1.3.4-beta
-
 require (
-	github.com/cespare/xxhash/v2 v2.3.0 // indirect
-	github.com/dgraph-io/ristretto/v2 v2.3.0 // indirect
-	github.com/dustin/go-humanize v1.0.1 // indirect
-	github.com/google/uuid v1.6.0 // indirect
-	github.com/jackc/pgpassfile v1.0.0 // indirect
-	github.com/jackc/pgservicefile v0.0.0-20240606120523-5a60cdf6a761 // indirect
-	github.com/jackc/pgx/v5 v5.7.5 // indirect
-	github.com/jackc/puddle/v2 v2.2.2 // indirect
-	github.com/jinzhu/inflection v1.0.0 // indirect
-	github.com/jinzhu/now v1.1.5 // indirect
-	github.com/lib/pq v1.10.9 // indirect
-	github.com/mattn/go-sqlite3 v1.14.32 // indirect
-	golang.org/x/crypto v0.41.0 // indirect
-	golang.org/x/sync v0.16.0 // indirect
-	golang.org/x/sys v0.35.0 // indirect
-	golang.org/x/text v0.28.0 // indirect
-	gorm.io/driver/postgres v1.6.0 // indirect
-	gorm.io/driver/sqlite v1.6.0 // indirect
-	gorm.io/gorm v1.30.2 // indirect
+	github.com/Rhaqim/buckt v1.0.2
+	github.com/Rhaqim/buckt/client/web v1.0.2
+	github.com/Rhaqim/buckt/cloud/aws v1.0.2
+	github.com/Rhaqim/buckt/cloud/azure v1.0.2
+	github.com/Rhaqim/buckt/cloud/gcp v1.0.2
 )
 
 // add module from src folder
 replace github.com/Rhaqim/buckt => ../
+
+replace github.com/Rhaqim/buckt/client/web => ../client/web
+
+replace github.com/Rhaqim/buckt/cloud/aws => ../cloud/aws
+
+replace github.com/Rhaqim/buckt/cloud/azure => ../cloud/azure
+
+replace github.com/Rhaqim/buckt/cloud/gcp => ../cloud/gcp
