+++ internal/backend/localfs.go
@@ -1,146 +1,125 @@
 package backend
 
 import (
-	"errors"
-	"fmt"
 	"io"
 	"os"
 	"path/filepath"
 
 	"github.com/Rhaqim/buckt/internal/domain"
+	"github.com/Rhaqim/buckt/internal/model"
 	"github.com/Rhaqim/buckt/pkg/logger"
 	"golang.org/x/sync/singleflight"
 )
 
-type FileSystemService struct {
+type LocalFileSystemService struct {
 	*logger.BucktLogger
-
 	MediaDir string
-
-	g     singleflight.Group
-	cache domain.LRUCache
+	g        singleflight.Group
+	cache    domain.LRUCache
 }
 
-func NewFileSystemService(bucktLogger *logger.BucktLogger, medaiDir string, cache domain.LRUCache) domain.FileSystemService {
-	bucktLogger.Info("🚀 Initialising file system services")
-	return &FileSystemService{
+func NewLocalFileSystemService(bucktLogger *logger.BucktLogger, mediaDir string, cache domain.LRUCache) domain.FileBackend {
+	bucktLogger.Info("🚀 Initialising local file system backend")
+	return &LocalFileSystemService{
 		BucktLogger: bucktLogger,
-		MediaDir:    medaiDir,
-
-		g:     singleflight.Group{},
-		cache: cache,
+		MediaDir:    mediaDir,
+		g:           singleflight.Group{},
+		cache:       cache,
 	}
 }
 
-func (bfs *FileSystemService) FSValidatePath(path string) (string, error) {
-	filePath := filepath.Join(bfs.MediaDir, path)
-
-	if _, err := os.Stat(filePath); err != nil {
-		return "", bfs.WrapError("failed to validate file path", err)
-	}
-
-	return filePath, nil
+func (bfs *LocalFileSystemService) resolve(path string) string {
+	return filepath.Join(bfs.MediaDir, path)
 }
 
-func (bfs *FileSystemService) FSWriteFile(filePath string, file []byte) error {
-	// File system path
-	filePath = filepath.Join(bfs.MediaDir, filePath)
-
-	// Save the file to the file system
-	if err := os.MkdirAll(filepath.Dir(filePath), os.ModePerm); err != nil {
+// Put writes/overwrites a file.
+func (bfs *LocalFileSystemService) Put(path string, data []byte) error {
+	filePath := bfs.resolve(path)
+	if err := os.MkdirAll(filepath.Dir(filePath), 0755); err != nil {
 		return bfs.WrapError("failed to create directory", err)
 	}
-
-	if err := os.WriteFile(filePath, file, 0644); err != nil {
-		return bfs.WrapError("failed to write file", err)
-	}
-
-	return nil
+	return os.WriteFile(filePath, data, 0644)
 }
 
-func (bfs *FileSystemService) FSGetFile(path string) ([]byte, error) {
-	filePath, err := bfs.FSValidatePath(path)
-	if err != nil {
-		return nil, err
-	}
+// Get reads the entire file into memory.
+func (bfs *LocalFileSystemService) Get(path string) ([]byte, error) {
+	filePath := bfs.resolve(path)
 
-	if file, ok := bfs.cache.Get(filePath); ok {
-		return file, nil
+	// check cache first
+	if data, ok := bfs.cache.Get(filePath); ok {
+		return data, nil
 	}
 
 	result, err, _ := bfs.g.Do(filePath, func() (any, error) {
 		return os.ReadFile(filePath)
 	})
-
 	if err != nil {
 		return nil, bfs.WrapError("failed to read file", err)
 	}
 
-	// check what type result is
-	if _, ok := result.([]byte); !ok {
-		return nil, bfs.WrapError(fmt.Sprintf("failed to read file: expected []byte but got %T", result), errors.New("unexpected type"))
-	}
-	// bfs.cache.Add(filePath, result.([]byte))
-	bfs.cache.Add(filePath, result.([]byte))
-
-	return result.([]byte), nil
+	bytes := result.([]byte)
+	bfs.cache.Add(filePath, bytes)
+	return bytes, nil
 }
 
-func (bfs *FileSystemService) FSGetFileStream(path string) (io.ReadCloser, error) {
-	filePath, err := bfs.FSValidatePath(path)
-	if err != nil {
-		return nil, err
-	}
-
-	file, err := os.Open(filePath)
-	if err != nil {
-		return nil, err
-	}
-
-	return file, nil // Caller should close the file after reading
+// Stream returns a file stream (caller must Close).
+func (bfs *LocalFileSystemService) Stream(path string) (io.ReadCloser, error) {
+	filePath := bfs.resolve(path)
+	return os.Open(filePath)
 }
 
-func (bfs *FileSystemService) FSUpdateFile(oldPath, newPath string) error {
-	oldFilePath, err := bfs.FSValidatePath(oldPath)
-	if err != nil {
-		return err
+// Delete removes a file.
+func (bfs *LocalFileSystemService) Delete(path string) error {
+	filePath := bfs.resolve(path)
+	if err := os.Remove(filePath); err != nil {
+		return bfs.WrapError("failed to delete file", err)
 	}
+	return nil
+}
 
-	newFilePath, err := bfs.FSValidatePath(newPath)
-	if err != nil {
-		return err
+// Exists checks if a file exists.
+func (bfs *LocalFileSystemService) Exists(path string) (bool, error) {
+	filePath := bfs.resolve(path)
+	_, err := os.Stat(filePath)
+	if err == nil {
+		return true, nil
 	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return false, bfs.WrapError("failed to stat file", err)
+}
+
+// Stat returns metadata.
+func (bfs *LocalFileSystemService) Stat(path string) (*model.FileInfo, error) {
+	filePath := bfs.resolve(path)
+	info, err := os.Stat(filePath)
+	if err != nil {
+		return nil, bfs.WrapError("failed to stat file", err)
+	}
+	return &model.FileInfo{
+		Size:         info.Size(),
+		LastModified: info.ModTime(),
+		ETag:         "",                         // Not applicable locally
+		ContentType:  "application/octet-stream", // crude default
+	}, nil
+}
+
+func (bfs *LocalFileSystemService) DeleteFolder(path string) error {
+	dirPath := bfs.resolve(path)
+	return os.RemoveAll(dirPath)
+}
+
+func (bfs *LocalFileSystemService) Move(oldPath, newPath string) error {
+	oldFilePath := filepath.Join(bfs.MediaDir, oldPath)
+	newFilePath := filepath.Join(bfs.MediaDir, newPath)
 
 	if err := os.MkdirAll(filepath.Dir(newFilePath), 0755); err != nil {
 		return bfs.WrapError("failed to create directory", err)
 	}
 
 	if err := os.Rename(oldFilePath, newFilePath); err != nil {
-		return bfs.WrapError("failed to update file", err)
+		return bfs.WrapError("failed to move file", err)
 	}
-
-	return nil
-}
-
-func (bfs *FileSystemService) FSDeleteFile(folderPath string) error {
-	filePath, err := bfs.FSValidatePath(folderPath)
-	if err != nil {
-		return err
-	}
-
-	if err := os.Remove(filePath); err != nil {
-		return bfs.WrapError("failed to delete file", err)
-	}
-
-	return nil
-}
-
-func (bfs *FileSystemService) FSDeleteFolder(folderPath string) error {
-	folderPath = filepath.Join(bfs.MediaDir, folderPath)
-
-	if err := os.RemoveAll(folderPath); err != nil {
-		return bfs.WrapError("failed to delete folder", err)
-	}
-
 	return nil
 }
