+++ internal/backend/localfs_old_test.go
@@ -5,21 +5,44 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/Rhaqim/buckt/internal/mocks"
 	"github.com/Rhaqim/buckt/pkg/logger"
 	"github.com/stretchr/testify/assert"
 )
 
-func setupFSV2Test() (*LocalFileSystemService, string) {
+type MockLRUCache struct{}
+
+func (m *MockLRUCache) Close()                            {}
+func (M *MockLRUCache) Get(key string) ([]byte, bool)     { return nil, false }
+func (M *MockLRUCache) Add(key string, value []byte) bool { return true }
+func (M *MockLRUCache) Hits() uint64                      { return 0 }
+func (M *MockLRUCache) Misses() uint64                    { return 0 }
+
+func setupFSTest() (*FileSystemService, string) {
 	log := logger.NewLogger("", true, false)
 	mediaDir := os.TempDir()
-	cache := new(mocks.LRUCache)
-	bfs := NewLocalFileSystemService(log, mediaDir, cache).(*LocalFileSystemService)
+	cache := &MockLRUCache{}
+	bfs := NewFileSystemService(log, mediaDir, cache).(*FileSystemService)
 	return bfs, mediaDir
 }
 
-func TestFSV2Put(t *testing.T) {
-	bfs, mediaDir := setupFSV2Test()
+func TestFSValidatePath(t *testing.T) {
+	bfs, mediaDir := setupFSTest()
+	testPath := "testfile.txt"
+	expectedPath := filepath.Join(mediaDir, testPath)
+
+	// Create a test file
+	_, err := os.Create(expectedPath)
+	assert.NoError(t, err)
+	defer os.Remove(expectedPath)
+
+	// Validate path
+	validatedPath, err := bfs.FSValidatePath(testPath)
+	assert.NoError(t, err)
+	assert.Equal(t, expectedPath, validatedPath)
+}
+
+func TestFSWriteFile(t *testing.T) {
+	bfs, mediaDir := setupFSTest()
 	testPath := "testfile.txt"
 	testContent := []byte("Hello, World!")
 	expectedPath := filepath.Join(mediaDir, testPath)
@@ -29,7 +52,7 @@ func TestFSV2Put(t *testing.T) {
 	assert.NoError(t, err)
 
 	// Write file
-	err = bfs.Put(testPath, testContent)
+	err = bfs.FSWriteFile(testPath, testContent)
 	assert.NoError(t, err)
 	defer os.Remove(expectedPath)
 
@@ -39,8 +62,8 @@ func TestFSV2Put(t *testing.T) {
 	assert.Equal(t, testContent, content)
 }
 
-func TestFSV2GetFile(t *testing.T) {
-	bfs, mediaDir := setupFSV2Test()
+func TestFSGetFile(t *testing.T) {
+	bfs, mediaDir := setupFSTest()
 	testPath := "testfile.txt"
 	testContent := []byte("Hello, World!")
 	expectedPath := filepath.Join(mediaDir, testPath)
@@ -55,13 +78,13 @@ func TestFSV2GetFile(t *testing.T) {
 	defer os.Remove(expectedPath)
 
 	// Get file
-	content, err := bfs.Get(testPath)
+	content, err := bfs.FSGetFile(testPath)
 	assert.NoError(t, err)
 	assert.Equal(t, testContent, content)
 }
 
-func TestFSV2Move(t *testing.T) {
-	bfs, mediaDir := setupFSV2Test()
+func TestFSUpdateFile(t *testing.T) {
+	bfs, mediaDir := setupFSTest()
 	oldPath := "oldfile.txt"
 	newPath := "newfile.txt"
 	testContent := []byte("Hello, World!")
@@ -82,8 +105,8 @@ func TestFSV2Move(t *testing.T) {
 	defer os.Remove(oldFilePath)
 	defer os.Remove(newFilePath)
 
-	// Move file
-	err = bfs.Move(oldPath, newPath)
+	// Update file
+	err = bfs.FSUpdateFile(oldPath, newPath)
 	assert.NoError(t, err)
 
 	// Validate old file does not exist
@@ -96,8 +119,8 @@ func TestFSV2Move(t *testing.T) {
 	assert.Equal(t, testContent, content)
 }
 
-func TestFSV2DeleteFile(t *testing.T) {
-	bfs, mediaDir := setupFSV2Test()
+func TestFSDeleteFile(t *testing.T) {
+	bfs, mediaDir := setupFSTest()
 	testPath := "testfile.txt"
 	expectedPath := filepath.Join(mediaDir, testPath)
 
@@ -106,15 +129,15 @@ func TestFSV2DeleteFile(t *testing.T) {
 	assert.NoError(t, err)
 
 	// Delete file
-	err = bfs.Delete(testPath)
+	err = bfs.FSDeleteFile(testPath)
 	assert.NoError(t, err)
 
 	// Validate file does not exist
 	_, err = os.Stat(expectedPath)
 	assert.True(t, os.IsNotExist(err))
 }
-func TestFSV2DeleteFolder(t *testing.T) {
-	bfs, mediaDir := setupFSV2Test()
+func TestFSDeleteFolder(t *testing.T) {
+	bfs, mediaDir := setupFSTest()
 	testFolderPath := "testfolder"
 	expectedFolderPath := filepath.Join(mediaDir, testFolderPath)
 
@@ -128,7 +151,7 @@ func TestFSV2DeleteFolder(t *testing.T) {
 	assert.NoError(t, err)
 
 	// Delete folder
-	err = bfs.DeleteFolder(testFolderPath)
+	err = bfs.FSDeleteFolder(testFolderPath)
 	assert.NoError(t, err)
 
 	// Validate folder does not exist
