+++ internal/backend/localfs_test.go
@@ -5,44 +5,21 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/Rhaqim/buckt/internal/mocks"
 	"github.com/Rhaqim/buckt/pkg/logger"
 	"github.com/stretchr/testify/assert"
 )
 
-type MockLRUCache struct{}
-
-func (m *MockLRUCache) Close()                            {}
-func (M *MockLRUCache) Get(key string) ([]byte, bool)     { return nil, false }
-func (M *MockLRUCache) Add(key string, value []byte) bool { return true }
-func (M *MockLRUCache) Hits() uint64                      { return 0 }
-func (M *MockLRUCache) Misses() uint64                    { return 0 }
-
-func setupFSTest() (*FileSystemService, string) {
+func setupFSV2Test() (*LocalFileSystemService, string) {
 	log := logger.NewLogger("", true, false)
 	mediaDir := os.TempDir()
-	cache := &MockLRUCache{}
-	bfs := NewFileSystemService(log, mediaDir, cache).(*FileSystemService)
+	cache := new(mocks.LRUCache)
+	bfs := NewLocalFileSystemService(log, mediaDir, cache).(*LocalFileSystemService)
 	return bfs, mediaDir
 }
 
-func TestFSValidatePath(t *testing.T) {
-	bfs, mediaDir := setupFSTest()
-	testPath := "testfile.txt"
-	expectedPath := filepath.Join(mediaDir, testPath)
-
-	// Create a test file
-	_, err := os.Create(expectedPath)
-	assert.NoError(t, err)
-	defer os.Remove(expectedPath)
-
-	// Validate path
-	validatedPath, err := bfs.FSValidatePath(testPath)
-	assert.NoError(t, err)
-	assert.Equal(t, expectedPath, validatedPath)
-}
-
-func TestFSWriteFile(t *testing.T) {
-	bfs, mediaDir := setupFSTest()
+func TestFSV2Put(t *testing.T) {
+	bfs, mediaDir := setupFSV2Test()
 	testPath := "testfile.txt"
 	testContent := []byte("Hello, World!")
 	expectedPath := filepath.Join(mediaDir, testPath)
@@ -52,7 +29,7 @@ func TestFSWriteFile(t *testing.T) {
 	assert.NoError(t, err)
 
 	// Write file
-	err = bfs.FSWriteFile(testPath, testContent)
+	err = bfs.Put(testPath, testContent)
 	assert.NoError(t, err)
 	defer os.Remove(expectedPath)
 
@@ -62,8 +39,8 @@ func TestFSWriteFile(t *testing.T) {
 	assert.Equal(t, testContent, content)
 }
 
-func TestFSGetFile(t *testing.T) {
-	bfs, mediaDir := setupFSTest()
+func TestFSV2GetFile(t *testing.T) {
+	bfs, mediaDir := setupFSV2Test()
 	testPath := "testfile.txt"
 	testContent := []byte("Hello, World!")
 	expectedPath := filepath.Join(mediaDir, testPath)
@@ -78,13 +55,13 @@ func TestFSGetFile(t *testing.T) {
 	defer os.Remove(expectedPath)
 
 	// Get file
-	content, err := bfs.FSGetFile(testPath)
+	content, err := bfs.Get(testPath)
 	assert.NoError(t, err)
 	assert.Equal(t, testContent, content)
 }
 
-func TestFSUpdateFile(t *testing.T) {
-	bfs, mediaDir := setupFSTest()
+func TestFSV2Move(t *testing.T) {
+	bfs, mediaDir := setupFSV2Test()
 	oldPath := "oldfile.txt"
 	newPath := "newfile.txt"
 	testContent := []byte("Hello, World!")
@@ -105,8 +82,8 @@ func TestFSUpdateFile(t *testing.T) {
 	defer os.Remove(oldFilePath)
 	defer os.Remove(newFilePath)
 
-	// Update file
-	err = bfs.FSUpdateFile(oldPath, newPath)
+	// Move file
+	err = bfs.Move(oldPath, newPath)
 	assert.NoError(t, err)
 
 	// Validate old file does not exist
@@ -119,8 +96,8 @@ func TestFSUpdateFile(t *testing.T) {
 	assert.Equal(t, testContent, content)
 }
 
-func TestFSDeleteFile(t *testing.T) {
-	bfs, mediaDir := setupFSTest()
+func TestFSV2DeleteFile(t *testing.T) {
+	bfs, mediaDir := setupFSV2Test()
 	testPath := "testfile.txt"
 	expectedPath := filepath.Join(mediaDir, testPath)
 
@@ -129,15 +106,15 @@ func TestFSDeleteFile(t *testing.T) {
 	assert.NoError(t, err)
 
 	// Delete file
-	err = bfs.FSDeleteFile(testPath)
+	err = bfs.Delete(testPath)
 	assert.NoError(t, err)
 
 	// Validate file does not exist
 	_, err = os.Stat(expectedPath)
 	assert.True(t, os.IsNotExist(err))
 }
-func TestFSDeleteFolder(t *testing.T) {
-	bfs, mediaDir := setupFSTest()
+func TestFSV2DeleteFolder(t *testing.T) {
+	bfs, mediaDir := setupFSV2Test()
 	testFolderPath := "testfolder"
 	expectedFolderPath := filepath.Join(mediaDir, testFolderPath)
 
@@ -151,7 +128,7 @@ func TestFSDeleteFolder(t *testing.T) {
 	assert.NoError(t, err)
 
 	// Delete folder
-	err = bfs.FSDeleteFolder(testFolderPath)
+	err = bfs.DeleteFolder(testFolderPath)
 	assert.NoError(t, err)
 
 	// Validate folder does not exist
